+++ doc/reference/CHANGELOG.md
@@ -7,6 +7,9 @@ All notable changes to this project will be documented in this file.
 ### Added
 ### Removed
 ### Changed
+
+- zookeeper - increase certificate validity duration to 25 years to avoid renewal burden
+
 ### Fixed
 ### Security